requirements.txt
lxml==3.6.4
scrapy==1.2.1
sendgrid==3.6.0
pyyaml==3.12
tinydb==3.2.1
